Update functions.py
Adds the additional attributes "Time" and "Update_Mac" to the pet sensor.

"Time" ist the timestamp of the last update (can be used to trigger automations on change), "Update_Mac" is the MAC address of the device which did the last change.

With these information you can track the location of the pet in HA in setups with multiple flaps (i.e. one from the house into a screened area, another one from the screened area into the garden etc.)

Generation of pethubconfig.json from cloud adds this to pethubconfig as well.