Watchdog for Hub "Message of death"
This commit adds a watchdog to solve issue #11 by implementing my idea #17.

Even without pet action or flap status messages, the hub sends a status telegram at least every hour.

The watchdog checks every minute the time since the last message from the hub. If PHL did not see a message for at least 70 minutes, it assumes that the hub is trapped in its "loop of death" described in #11 and it'll stop Mosquitto, delete the persistent data file and restart Mosquitto. This usually solves the loop of death and the Hub will reconnect.

If the Hub is disconnected from the network or powered down, the Watchdog will also do the Mosquitto action every 70 minutes, but this won't harm anything.

Although the watchdog check runs every minute, a status is written into the log file as DEBUG message every 5 minutes to not flood the logfile.
When the watchdog triggers the action, an INFO message is written into the log file.

To function properly, these prerequsites must be fulfilled:
- The local Mosquitto for PHL must run via SYSTEMCTL
- The persistent Mosquitto data must reside in /var/lib/mosquitto/mosquitto.db (the standard location)
- Musquitto on the PHL server *must not be used for anything else but PHL* with an upstream connection to the main Moquitto broker. Otherwise, when the Watchdog triggers persistent data for other applications would also be deleted

If Mosquitto does not run via SYSTEMCTL or if the persistend data location is different, the command in frontend.py, line 315 must be modofied.